# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Data Collection Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets/versions/{dataset_version_id}/rows": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Dataset Rows",
        "description": "One page of a dataset version's rows, resumably.\n\n`cursor` is the `nextCursor` of the previous page and nothing else — not a\nrow number, not an offset. A published dataset version is immutable (§3.1),\nso the ordering cannot shift under a device that paused for a day, and a\nresumed transfer re-reads nothing.\n\n`nextCursor` is null on the last page. That is how a device knows it has the\n**whole** list rather than most of one, which matters more here than\nanywhere else in this API: a village list that stopped two thirds of the way\nthrough is a list an enumerator can search, scroll and choose from, and\nnothing about it looks wrong.\n\nImmutable, so a response may be cached forever.",
        "operationId": "dataset_rows_api_v1_datasets_versions__dataset_version_id__rows_get",
        "parameters": [
          {
            "name": "dataset_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Dataset Version Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 2000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetRowsPage"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/forms/versions/{form_version_id}": {
      "get": {
        "tags": [
          "forms"
        ],
        "summary": "Get Form Version",
        "description": "One published version and its Form IR.\n\nThe second half of form delivery (sync §5). `GET /sync/pull?scope=forms`\ntells a device which versions its environment runs and what they hash to;\nthis hands over a document the device does not already hold. Splitting the\ntwo is what keeps a sync cheap: the manifest is a few hundred bytes and\ntravels on every pull, the IR is tens of kilobytes and travels once.\n\nA published version is immutable (specs/erd-v0.1.md §4), so the response for\na given id can never change and a client may cache it forever.",
        "operationId": "get_form_version_api_v1_forms_versions__form_version_id__get",
        "parameters": [
          {
            "name": "form_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Form Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormVersionDocument"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sync/push": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Push",
        "description": "Accept a batch of operations, and the content keys they are encrypted to.\n\nEverything — content keys, op inserts, tombstones, state folds, outbox\nevents — commits in one transaction, so a key and the first ops it encrypts\narrive together or not at all. Replay is idempotent: an op the server\nalready has is reported accepted without being written again.",
        "operationId": "push_api_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/pull": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Pull",
        "description": "Resume the op and tombstone streams from a cursor, and optionally list\nthe form versions this device should be running.\n\n`scope` is the comma-separated list of spec §5: `assignments`, `forms`,\n`datasets`. `assignments` is accepted and ignored, so a newer client asking\nfor all three still works against this server rather than failing on an\nunknown word.\n\n`scope=datasets` returns the dataset versions the form versions deployed to\nthis device were **published against** (`form_version_dataset`), not the\ndatasets its project happens to own. The pinning that lets an answer be\nexplained later is the same thing that decides which rows travel. Like the\nform manifest it names versions and checksums, never rows: those are fetched\nonce per version, paged, from `GET /datasets/versions/{id}/rows`.\n\n`datasets` is **null** when nothing was asked or the device is unknown, and\n`[]` when the answer is genuinely \"none\" — a device must be able to tell an\nunanswered question from an answer, or it will delete a village list because\nit synced against an older server.\n\n`scope=forms` needs `deviceId`, because deployment is per environment\n(`form_deployment`): a device is told about the versions deployed to its own\nenvironment, never everything the project has published. The manifest names\nversions and their checksums, not their IR — the documents are fetched one\nat a time from `GET /forms/versions/{formVersionId}`, so a device that\nalready holds a version spends nothing re-reading it.\n\nUnlike the op stream, the manifest is a complete statement rather than a\ndelta: it is how a device notices a version has been *withdrawn*, which no\nstream of additions could tell it.",
        "operationId": "pull_api_v1_sync_pull_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "name": "deviceId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deviceid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentKeyIn": {
        "properties": {
          "contentKeyId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Contentkeyid"
          },
          "submissionId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Submissionid"
          },
          "deviceId": {
            "type": "string",
            "minLength": 1,
            "title": "Deviceid"
          },
          "wraps": {
            "items": {
              "$ref": "#/components/schemas/WrappedKeyIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Wraps"
          }
        },
        "type": "object",
        "required": [
          "contentKeyId",
          "submissionId",
          "deviceId",
          "wraps"
        ],
        "title": "ContentKeyIn",
        "description": "A device's content key for one submission, wrapped to every recipient.\n\nThe key material itself is never here — only copies the server cannot open."
      },
...
      "DatasetRowsPage": {
        "properties": {
          "datasetVersionId": {
            "type": "string",
            "title": "Datasetversionid"
          },
          "rows": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "hasMore": {
            "type": "boolean",
            "title": "Hasmore"
          }
        },
        "type": "object",
        "required": [
          "datasetVersionId",
          "rows",
          "nextCursor",
          "hasMore"
        ],
        "title": "DatasetRowsPage",
        "description": "One page of a dataset version's rows (sync §5).\n\nPaged because the first sync is the hard case and cannot be one response:\na transfer that cannot resume is a transfer that never finishes on the\nconnections this product exists for.\n\nA published version is immutable, so this page can be cached forever and a\ndevice that paused for a day resumes into the same ordering it left."
      },
      "DeployedDatasetVersion": {
        "properties": {
          "formVersionId": {
            "type": "string",
            "title": "Formversionid"
          },
          "datasetKey": {
            "type": "string",
            "title": "Datasetkey"
          },
          "datasetVersionId": {
            "type": "string",
            "title": "Datasetversionid"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "rowCount": {
            "type": "integer",
            "title": "Rowcount"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "filterColumns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filtercolumns"
          }
        },
        "type": "object",
        "required": [
          "formVersionId",
          "datasetKey",
          "datasetVersionId",
          "version",
          "rowCount",
          "checksum"
        ],
        "title": "DeployedDatasetVersion",
        "description": "One entry of a device's dataset manifest (sync §5, `scope=datasets`).\n\nDeliberately not the rows. A village list is megabytes and a manifest\ntravels on every sync; the rows are fetched once per version from\n`GET /datasets/versions/{id}/rows`, the same split that keeps a form\nmanifest to a few hundred bytes.\n\n`formVersionId` is on every entry rather than implied, because the pin is\nper form version: two versions of a form can be deployed at once — an\nenumerator holding a v2 draft the morning v3 lands — and they may name\ndifferent versions of the same list. A manifest keyed only by dataset would\nhave to choose between them, which is the choice §3.2 exists to remove."
      },
      "DeployedFormVersion": {
        "properties": {
          "formVersionId": {
            "type": "string",
            "title": "Formversionid"
          },
          "formId": {
            "type": "string",
            "title": "Formid"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "irChecksum": {
            "type": "string",
            "title": "Irchecksum"
          },
          "deployedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Deployedat"
          }
        },
        "type": "object",
        "required": [
          "formVersionId",
          "formId",
          "version",
          "title",
          "irChecksum",
          "deployedAt"
        ],
        "title": "DeployedFormVersion",
        "description": "One entry in a device's form manifest (sync §5, `scope=forms`).\n\nDeliberately not the IR. A 52-question form is tens of kilobytes, and a\ndevice re-syncs on whatever connection it has; sending every document on\nevery pull would spend exactly the bandwidth this protocol exists to\nconserve. The manifest says what exists and what it hashes to, and the\ndevice fetches only the versions it does not already hold — the same shape\nresumable upload uses, where the server states what it has and the client\nsends the rest."
      },
...
      "FormVersionDocument": {
        "properties": {
          "formVersionId": {
            "type": "string",
            "title": "Formversionid"
          },
          "formId": {
            "type": "string",
            "title": "Formid"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "irChecksum": {
            "type": "string",
            "title": "Irchecksum"
          },
          "publishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishedat"
          },
          "form": {
            "additionalProperties": true,
            "type": "object",
            "title": "Form"
          }
        },
        "type": "object",
        "required": [
          "formVersionId",
          "formId",
          "version",
          "title",
          "irChecksum",
          "publishedAt",
          "form"
        ],
        "title": "FormVersionDocument",
        "description": "One published version and its Form IR (sync §5).\n\nWhat a device fetches once the manifest names a version it does not hold.\nImmutable: the id addresses a row that can never be rewritten\n(specs/erd-v0.1.md §4), so a client may cache it forever."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "MessageError",
        "description": "The body of a refusal that carries prose and nothing to branch on.\n\n`{\"detail\": \"submission not found\"}`. Used where there is exactly one way\nto fail and the status code already says which: a 404 on a read endpoint.\nAnything a client must branch on gets a `reason` field instead — see\n`DeviceRegisterError` and `ProjectKeyError` in `modules/projects/schemas.py`."
      },
      "OpKind": {
        "type": "string",
        "enum": [
          "set",
          "unset",
          "repeat_add",
          "repeat_delete",
          "finalize",
          "reopen"
        ]
      },
...
      "PullResponse": {
        "properties": {
          "ops": {
            "items": {
              "$ref": "#/components/schemas/PulledOp"
            },
            "type": "array",
            "title": "Ops"
          },
          "tombstones": {
            "items": {
              "$ref": "#/components/schemas/PulledTombstone"
            },
            "type": "array",
            "title": "Tombstones"
          },
          "forms": {
            "items": {
              "$ref": "#/components/schemas/DeployedFormVersion"
            },
            "type": "array",
            "title": "Forms"
          },
          "datasets": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DeployedDatasetVersion"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datasets"
          },
          "nextCursor": {
            "type": "integer",
            "title": "Nextcursor"
          },
          "hasMore": {
            "type": "boolean",
            "title": "Hasmore"
          }
        },
        "type": "object",
        "required": [
          "ops",
          "tombstones",
          "forms",
          "nextCursor",
          "hasMore"
        ],
        "title": "PullResponse"
      },
      "PulledOp": {
        "properties": {
          "opId": {
            "type": "string",
            "title": "Opid"
          },
          "submissionId": {
            "type": "string",
            "title": "Submissionid"
          },
          "formId": {
            "type": "string",
            "title": "Formid"
          },
          "formVersion": {
            "type": "integer",
            "title": "Formversion"
          },
          "kind": {
            "$ref": "#/components/schemas/OpKind"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "value": {
            "title": "Value"
          },
          "valueCiphertext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valueciphertext"
          },
          "contentKeyId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentkeyid"
          },
          "nonce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce"
          },
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "actorId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorid"
          },
          "counter": {
            "type": "integer",
            "title": "Counter"
          },
          "wallClock": {
            "type": "string",
            "format": "date-time",
            "title": "Wallclock"
          },
          "serverSeq": {
            "type": "integer",
            "title": "Serverseq"
          }
        },
        "type": "object",
        "required": [
          "opId",
          "submissionId",
          "formId",
          "formVersion",
          "kind",
          "path",
          "value",
          "valueCiphertext",
          "contentKeyId",
          "nonce",
          "deviceId",
          "actorId",
          "counter",
          "wallClock",
          "serverSeq"
        ],
        "title": "PulledOp"
      },
      "PulledTombstone": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subjectType": {
            "$ref": "#/components/schemas/TombstoneSubject"
          },
          "subjectId": {
            "type": "string",
            "title": "Subjectid"
          },
          "submissionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submissionid"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "deviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviceid"
          },
          "counter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "serverSeq": {
            "type": "integer",
            "title": "Serverseq"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subjectType",
          "subjectId",
          "submissionId",
          "path",
          "deviceId",
          "counter",
          "createdAt",
          "expiresAt",
          "serverSeq"
        ],
        "title": "PulledTombstone"
      },
      "PushRequest": {
        "properties": {
          "deviceId": {
            "type": "string",
            "minLength": 1,
            "title": "Deviceid"
          },
          "ops": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Ops"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/ContentKeyIn"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Keys"
          }
        },
        "type": "object",
        "required": [
          "deviceId",
          "ops"
        ],
        "title": "PushRequest"
      },
      "PushResponse": {
        "properties": {
          "accepted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedOp"
            },
            "type": "array",
            "title": "Rejected"
          },
          "serverCursor": {
            "type": "integer",
            "title": "Servercursor"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "rejected",
          "serverCursor"
        ],
        "title": "PushResponse"
      },
...
      "RejectReason": {
        "type": "string",
        "enum": [
          "unknown_form_version",
          "not_authorized",
          "submission_closed",
          "malformed",
          "unknown_content_key",
          "nonce_reused"
        ]
      },
      "RejectedOp": {
        "properties": {
          "opId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opid"
          },
          "reason": {
            "$ref": "#/components/schemas/RejectReason"
          }
        },
        "type": "object",
        "required": [
          "opId",
          "reason"
        ],
        "title": "RejectedOp"
      },
...
      "TombstoneSubject": {
        "type": "string",
        "enum": [
          "submission",
          "repeat_instance",
          "case",
          "entity",
          "media"
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WrappedKeyIn": {
        "properties": {
          "projectKeyId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Projectkeyid"
          },
          "ephemeralPublic": {
            "type": "string",
            "title": "Ephemeralpublic"
          },
          "nonce": {
            "type": "string",
            "title": "Nonce"
          },
          "wrappedKey": {
            "type": "string",
            "title": "Wrappedkey"
          }
        },
        "type": "object",
        "required": [
          "projectKeyId",
          "ephemeralPublic",
          "nonce",
          "wrappedKey"
        ],
        "title": "WrappedKeyIn",
        "description": "One content key wrapped to one recipient project key (envelope §4.3)."
      },
...
    }
  }
}